Fix bad behaviour when there are no event inside a quantum.
ThreadSafeEvents are posted directly after the next posted event.
If there is no event inside a quantum, quantum_notify will make SystemC sleep
and wait for QBOX to finish but post_a_transaction will wait for SystemC to post
the event. Just don't sleep in this case so if there are pending transactions
they will be triggered directly.

Signed-off-by: KONRAD Frederic <fred.konrad@greensocs.com>